Intermediate-LLMOPS/requirements.txt
fastapi
uvicorn
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface
pypdf
sentence-transformers
weaviate-client
llama-index-vector-stores-weaviate
